# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/trips": {
      "get": {
        "operationId": "listTrips",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "driverId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refueled",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "createTrip",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripInput"
              }
            }
          }
        }
      }
    },
...
    "/trips/{id}": {
      "put": {
        "operationId": "correctTrip",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/fuel": {
      "get": {
        "operationId": "getFuelSummary",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuelSummary"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "driverId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refueled",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TripInput": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "integer"
          },
          "vehicleId": {
            "type": "integer"
          },
          "initialKm": {
            "type": "number"
          },
          "finalKm": {
            "type": "number"
          },
          "origin": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "companions": {
            "type": "string"
          },
          "refueled": {
            "type": "boolean"
          },
          "liters": {
            "type": "number"
          },
          "fuelCost": {
            "type": "number"
          },
          "observations": {
            "type": "string"
          }
        },
        "required": [
          "driverId",
          "vehicleId",
          "initialKm",
          "finalKm",
          "origin",
          "destination",
          "refueled"
        ]
      },
...
      "CorrectionInput": {
        "type": "object",
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripInput"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "trip",
          "reason"
        ]
      },
...
      "FuelSummary": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer"
          },
          "plate": {
            "type": "string"
          },
          "liters": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "loads": {
            "type": "integer"
          },
          "averageLiters": {
            "type": "number"
          },
          "distance": {
            "type": "number"
          },
          "kmPerLiter": {
            "type": "number"
          }
        },
        "required": [
          "vehicleId",
          "plate",
          "liters",
          "cost",
          "loads",
          "averageLiters",
          "distance",
          "kmPerLiter"
        ]
      },
...
      "Trip": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "integer"
          },
          "vehicleId": {
            "type": "integer"
          },
          "initialKm": {
            "type": "number"
          },
          "finalKm": {
            "type": "number"
          },
          "origin": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "companions": {
            "type": "string"
          },
          "refueled": {
            "type": "boolean"
          },
          "liters": {
            "type": "number"
          },
          "fuelCost": {
            "type": "number"
          },
          "observations": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "at": {
            "type": "string"
          },
          "driverName": {
            "type": "string"
          },
          "plate": {
            "type": "string"
          },
          "distance": {
            "type": "number"
          },
          "corrected": {
            "type": "boolean"
          }
        },
        "required": [
          "driverId",
          "vehicleId",
          "initialKm",
          "finalKm",
          "origin",
          "destination",
          "companions",
          "refueled",
          "liters",
          "fuelCost",
          "observations",
          "id",
          "at",
          "driverName",
          "plate",
          "distance",
          "corrected"
        ]
      },
...
    }
  }
}