# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Trip {
  driverId: number;
  vehicleId: number;
  initialKm: number;
  finalKm: number;
  origin: string;
  destination: string;
  companions: string;
  refueled: boolean;
  liters: number;
  fuelCost: number;
  observations: string;
  id: number;
  at: string;
  driverName: string;
  plate: string;
  distance: number;
  corrected: boolean;
}